establish functions to prevent user to update and delete a goal that is not theirs

+++ backend/controllers/goalController.js
@@ -1,5 +1,6 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // desc Get goals
 // @route GET /api/goals
@@ -40,6 +41,20 @@ const udpateGoal = asyncHandler(async (req, res) => {
     throw new Error("Goal not found");
   }
 
+  const user = await User.findById(req.user.id)
+
+  // Check for user
+  if(!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  // Make sure the logged in user matches the goal user
+  if(goal.user.toString() !== user.id){
+    res.status(401)
+    throw new Error('User not authorized')
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -57,6 +72,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id)
+
+  // Check for user
+  if(!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  // Make sure the logged in user matches the goal user
+  if(goal.user.toString() !== user.id){
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   await Goal.deleteOne(goal)
